Mod min

+++ iconize/pages/events.py
@@ -10,20 +10,27 @@ from ..utils.img import make512
 from ..utils.fileOpe import saveFile, deleteFile
 import os
 
-bp = Blueprint('events', __name__, url_prefix='/posts/')
+bp = Blueprint("events", __name__, url_prefix="/posts/")
 
-
-@bp.route('/', methods=['GET'])
+@bp.route("/", methods=["GET"])
 def list_posts():
     num = 10
     list_of_posts = get_post_list(num)
     num_of_posts = list_of_posts.count()
-    return render_template('/events/index.html', posts=list_of_posts, num=num_of_posts)
+    return render_template("/events/index.html", posts=list_of_posts, num=num_of_posts)
 
+@bp.route("/<iD>/", methods=["GET"])
+def show_post(iD=None):
+    post = get_post(iD)
+    if post is None:
+        return render_template("events/error.html")
+    return render_template("/events/show_post.html", title=post.title, iD=iD,
+                           date=post.date, author=post.author, s_title=post.s_title,
+                           updated_date=str(post.updated_date), color=post.color)
 
-@bp.route('/create/', methods=['GET', 'POST'])
+@bp.route("/create/", methods=["GET", "POST"])
 def create_post():
-    if request.method == 'POST':
+    if request.method == "POST":
         try:
             title = request.form["title"]
             s_title = request.form["s_title"]
@@ -33,9 +40,9 @@ def create_post():
             returnValue = getHash(cleaned_html)
             iD = returnValue[0]
             token = returnValue[1]
-            add_post(iD, request.form['author'], cleaned_html,
+            add_post(iD, request.form["author"], cleaned_html,
                      title, s_title,
-                     request.form['color'], request.form['date'], token)
+                     request.form["color"], request.form["date"], token)
             if "iconfile" in request.files:
                 saveFile(icon=request.files["iconfile"], name=iD)
             res = make_response(jsonify(returnValue))
@@ -44,59 +51,56 @@ def create_post():
         except Exception as e:
             print(e)
             return "error"
-    return render_template('/events/content.html', access_type="Create")
+    return render_template("/events/content.html", access_type="Create")
 
 
-@bp.route('/<iD>/edit/', methods=['GET', 'POST'])
+@bp.route("/<iD>/edit/", methods=["GET", "POST"])
 def edit_post(iD):
     token_data = get_post(iD)
     if token_data is None:
-        return redirect('/')
+        return redirect("/")
     token = token_data.token
-    if request.method == 'POST':
+    if request.method == "POST":
+        token_data = get_post(iD)
+        if token_data == "" or token_data is None:
+            return "error"
+
+        token = token_data.token
+        if (request.form["token"] == "" or request.form["token"] is None or token != request.form["token"]) \
+                and (request.cookies.get(iD) is None or request.cookies.get(iD) != token):
+                return "error"
         try:
             title = request.form["title"]
-            s_title = request.form['s_title']
-            if s_title == '':
+            s_title = request.form["s_title"]
+            if s_title == "":
                 s_title = title
             if request.form["del"]:
                 deleteFile(iD)
             if "iconfile" in request.files:
                 saveFile(icon=request.files["iconfile"], name=iD)
             cleaned_html = cl(request.files["code"])
-            mod_post(iD, request.form['author'], cleaned_html,
+            mod_post(iD, request.form["author"], cleaned_html,
                      title, s_title,
-                     request.form['color'], request.form['date'])
-        except:
-            return 'error'
-        return 'success'
+                     request.form["color"], request.form["date"])
+        except Exception as e:
+            print(e)
+            return "error"
+        return "success"
 
-    if request.method == 'GET':
-        if (request.args.get('token') == '' or request.args.get('token') is None) and request.cookies.get(iD) is None:
-            return render_template('events/content.html', auth='missing')
-        elif request.args.get('token') != token and request.cookies.get(iD) != token:
-            return render_template('events/content.html', auth='incorrect')
+    if request.method == "GET":
+        if (request.args.get("token") == "" or request.args.get("token") is None) and request.cookies.get(iD) is None:
+            return render_template("events/content.html", auth="missing")
+        elif request.args.get("token") != token and request.cookies.get(iD) != token:
+            return render_template("events/content.html", auth="incorrect")
         else:
             post = get_post(iD)
-            return render_template('events/content.html', access_type="Edit",
-                                   author=post.author, auth='success', title=post.title,
+            return render_template("events/content.html", access_type="Edit",
+                                   author=post.author, auth="success", title=post.title,
                                    s_title=post.s_title, date=post.date, color=post.color)
 
 
-@bp.route('/<iD>/', methods=['GET'])
-def show_post(iD=None):
-    post = get_post(iD)
-    if post is None:
-        return render_template("events/error.html")
-    return render_template('/events/show_post.html', title=post.title, iD=iD,
-                           date=post.date, author=post.author, s_title=post.s_title,
-                           updated_date=str(post.updated_date), color=post.color)
-
-
-@bp.route("/<iD>/delete/", methods=["GET", "POST"])
+@bp.route("/<iD>/delete/", methods=["POST"])
 def delete(iD):
-    if request.method == "GET":
-        return redirect("/")
     if request.method == "POST":
         token_data = get_post(iD)
